requirements.txt
# (C) 2023 - Fateh Singh Khalsa
# Humidity Manager - Project Dependencies

python-dotenv
pysensorpush
pyvesync